+++ lib/mongoid-cached-json.rb
@@ -3,6 +3,5 @@ require 'mongoid'
 require 'active_support/concern'
 require 'mongoid-cached-json/version'
 require 'mongoid-cached-json/config'
-require 'mongoid-cached-json/downmark_it'
 require 'mongoid-cached-json/cached_json'