+++ spec/support/json_math.rb
@@ -1,0 +1,11 @@
+class JsonMath
+  include Mongoid::Document
+  include Mongoid::CachedJson
+
+  field :number
+
+  json_fields \
+    :number => {}
+    
+end
+