+++ HISTORY.md
@@ -4,4 +4,5 @@
 * 2011/11/15: Initial Implementation, [Aaron Windsor](http://github.com/aaw)
 * 2011/11/29: Added support for `:markdown`, [Frank Macreery](http://github.com/macreery)
 * 2012/02/17: Added `Mongoid::CachedJson::configure`, [Daniel Doubrovkine](http://github.com/dblock)
+* 2012/02/17: Added `Mongoid::CachedJson::transform`, [Daniel Doubrovkine](http://github.com/dblock)
 * TBD: Initial Public Release